requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==8.0.2
selenium==4.18.1
webdriver-manager==4.0.1
